backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "jsonschema>=4.26.0",
    "netmiko>=4.6.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "uvicorn[standard]>=0.44.0",
    "anthropic>=0.18.1",
    "websockets>=14.1",
    "pygnmi>=0.8.12",
    "prometheus_client>=0.21.1",
    "PyYAML>=6.0.1",
    "python-hcl2>=4.3.5",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
